Fix room layout save failing for stud-aligned positions
XCm/YCm snap to 0.8 cm (1 stud) so positions are floats like 1.6, 2.4, etc.
Changed int → double across PlacedTable model, document, DTOs, and export service.

+++ src/TroveKeep.Api/DTOs/Requests/RoomRequests.cs
@@ -6,7 +6,7 @@ public record UpdateRoomRequest(string Name, int WidthCm, int DepthCm, int Versi
 
 public record SaveRoomLayoutRequest(IEnumerable<PlacedTableRequest> Layout, IEnumerable<AggregateSelectionRequest> AggregateSelections, int Version = 0);
 
-public record PlacedTableRequest(Guid InstanceId, Guid TemplateId, int XCm, int YCm);
+public record PlacedTableRequest(Guid InstanceId, Guid TemplateId, double XCm, double YCm);
 
 public record AggregateSelectionRequest(string RepresentativeId, string BpKey);
