+++ src/TroveKeep.Repositories/Documents/PlacedTableDocument.cs
@@ -11,6 +11,6 @@ public class PlacedTableDocument
     [BsonGuidRepresentation(GuidRepresentation.Standard)]
     public Guid TemplateId { get; set; }
 
-    public int XCm { get; set; }
-    public int YCm { get; set; }
+    public double XCm { get; set; }
+    public double YCm { get; set; }
 }
